refactor: CLIエントリポイントを`bin/`に移動
ライブラリモジュールとエントリポイントが混在してややこしいため、
CLIから呼び出すだけのプログラムを`bin/`ディレクトリに分離しました。
