+++ plugins/kyosei/bin/detect-context.ts
@@ -3,12 +3,12 @@
  * SKILL.mdの埋め込みコマンドとして使用します。
  *
  * 使用例:
- *   node dist/detect-context.js "https://github.com/owner/repo/pull/123"
- *   node dist/detect-context.js
+ *   node dist/bin/detect-context.js "https://github.com/owner/repo/pull/123"
+ *   node dist/bin/detect-context.js
  */
 
 import process from "node:process";
-import { detectReviewContext } from "./context.js";
+import { detectReviewContext } from "../src/context.js";
 
 function main(): void {
   const argument = process.argv[2];
