+++ plugins/kyosei/bin/get-changeset.ts
@@ -3,14 +3,14 @@
  * SKILL.mdの埋め込みコマンドとして使用します。
  *
  * 使用例:
- *   node dist/src/get-changeset.js "https://github.com/owner/repo/pull/123"
- *   node dist/src/get-changeset.js
+ *   node dist/bin/get-changeset.js "https://github.com/owner/repo/pull/123"
+ *   node dist/bin/get-changeset.js
  */
 
 import process from "node:process";
-import { getChangeset } from "./changeset.js";
-import { createOctokitClient } from "./client.js";
-import { detectReviewContext } from "./context.js";
+import { getChangeset } from "../src/changeset.js";
+import { createOctokitClient } from "../src/client.js";
+import { detectReviewContext } from "../src/context.js";
 
 async function main(): Promise<void> {
   const argument = process.argv[2];
