+++ plugins/kyosei/skills/kyosei/SKILL.md
@@ -10,7 +10,7 @@ allowed-tools: Bash(node:*), Glob, Grep, Read, Task, mcp__github, mcp__github_in
 以下のコマンドでレビューコンテキストを判定します。
 結果はJSONで返されます。
 
-!`node ${CLAUDE_PLUGIN_ROOT}/dist/src/detect-context.js $ARGUMENTS`
+!`node ${CLAUDE_PLUGIN_ROOT}/dist/bin/detect-context.js $ARGUMENTS`
 
 ## JSONの解釈
 
@@ -35,7 +35,7 @@ JSONから主に以下の値を後続のコマンドで使用します:
 以下のコマンドで変更セットを取得します。
 結果はJSONで返されます。
 
-!`node ${CLAUDE_PLUGIN_ROOT}/dist/src/get-changeset.js $ARGUMENTS`
+!`node ${CLAUDE_PLUGIN_ROOT}/dist/bin/get-changeset.js $ARGUMENTS`
 
 ## JSONの解釈
